restructured design
