+++ src/App.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { BrowserRouter as Router, Route } from "react-router-dom";
-import Home from "../src/pages/Home/Home";
+import Home from "./pages/Home/Home";
 import Login from "../src/pages/Login/Login";
 import NotFoundPage from "../src/pages/NotFoundPage/NotFoundPage";
 import './scss/Main.scss';
