+++ src/scss/Abstracts/_Variables.scss
@@ -2,3 +2,4 @@
 
 $color-primary: #3F0E40;
 $color-white: #FFFFFF;
+$color-useronline: #3BCF3B;
