+++ src/scss/Components/_Home.scss
@@ -5,3 +5,8 @@
 body {
   margin: 0;
 }
+
+.user-online {
+  color: $color-useronline;
+  text-align: center;
+}
