+++ src/scss/Main.scss
@@ -8,5 +8,6 @@
 
 // Components
 @import "Components/Home";
+@import "Components/Chat";
 @import "Components/Sidebar";
 @import "Components/Channels";
